Rename registration rates to completness rates
