+++ packages/client/src/App.tsx
@@ -42,7 +42,7 @@ import { SelectInformant } from '@client/views/SelectInformant/SelectInformant'
 import { SelectVitalEvent } from '@client/views/SelectVitalEvent/SelectVitalEvent'
 import { SettingsPage } from '@client/views/Settings/SettingsPage'
 import { PerformanceHome } from '@client/views/SysAdmin/Performance/PerformanceHome'
-import { RegistrationRates } from '@client/views/SysAdmin/Performance/RegistrationRates'
+import { CompletenessRates } from '@client/views/SysAdmin/Performance/CompletenessRates'
 import { WorkflowStatus } from '@client/views/SysAdmin/Performance/WorkflowStatus'
 import { TeamSearch } from '@client/views/SysAdmin/Team/TeamSearch'
 import { CreateNewUser } from '@client/views/SysAdmin/Team/user/userCreation/CreateNewUser'
@@ -331,9 +331,9 @@ export class App extends React.Component<IAppProps> {
                                           <ProtectedRoute
                                             exact
                                             path={
-                                              routes.EVENT_REGISTRATION_RATES
+                                              routes.EVENT_COMPLETENESS_RATES
                                             }
-                                            component={RegistrationRates}
+                                            component={CompletenessRates}
                                           />
                                           <ProtectedRoute
                                             exact
