+++ packages/client/src/navigation/routes.ts
@@ -73,8 +73,8 @@ export const SYS_ADMIN_HOME = '/sys-admin-home'
 export const SYS_ADMIN_HOME_TAB = '/sys-admin-home/:tabId'
 
 export const PERFORMANCE_HOME = '/performance'
-export const EVENT_REGISTRATION_RATES =
-  '/performance/operations/regestrationRates/:eventType'
+export const EVENT_COMPLETENESS_RATES =
+  '/performance/operations/completenessRates/:eventType'
 export const WORKFLOW_STATUS = '/performance/operations/workflowStatus'
 export const PERFORMANCE_FIELD_AGENT_LIST = '/performance/field-agents'
