+++ packages/client/src/views/SysAdmin/Performance/WorkflowStatus.tsx
@@ -30,7 +30,7 @@ import { createNamesMap } from '@client/utils/data-formatting'
 import { EVENT_OPTIONS } from '@client/views/Performance/FieldAgentList'
 
 import { PerformanceSelect } from '@client/views/SysAdmin/Performance/PerformanceSelect'
-import { SORT_ORDER } from '@client/views/SysAdmin/Performance/reports/registrationRates/WithinTargetDaysTable'
+import { SORT_ORDER } from '@client/views/SysAdmin/Performance/reports/completenessRates/WithinTargetDaysTable'
 import { FilterContainer } from '@client/views/SysAdmin/Performance/utils'
 import {
   SysAdminContentWrapper,
